feat:a

+++ Controllers/ProductsController.cs
@@ -22,6 +22,7 @@ namespace Project_ShoeStore_Manager.Controllers
             var products = await context.Products
                          .Include(p => p.Brand)
                          .Include(p => p.Category)
+                         .Where(p => !p.IsDeleted)
                          .ToListAsync();
             return View(products);
         }
@@ -34,85 +35,66 @@ namespace Project_ShoeStore_Manager.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromForm] ProductDto productDto)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                using (var transaction = await context.Database.BeginTransactionAsync())
+                ViewData["BrandId"] = new SelectList(context.Brands, "BrandId", "BrandName");
+                ViewData["CategoryId"] = new SelectList(context.Categories, "CategoryId", "CategoryName");
+                return View(productDto);
+            }
+            using (var transaction = await context.Database.BeginTransactionAsync())
+            {
+                try
                 {
-                    try
+                    Product product = new Product();
                     {
-                        Product product = new Product();
+                        product.ProductName = productDto.ProductName;
+                        product.BrandId = productDto.BrandId;
+                        product.CategoryId = productDto.CategoryId;
+                        product.ProductDescription = productDto.ProductDescription;
+                        product.PurchasePrice = productDto.PurchasePrice;
+                        product.ProfitMargin = productDto.ProfitMargin;
+                    }
+
+                    product.ProductSizes = productDto.ProductSizes.Split(',', StringSplitOptions.RemoveEmptyEntries)
+                                                          .Select(size => new ProductSize()
+                                                          {
+                                                              SizeName = size.Trim()
+                                                          }).ToList();
+                    product.ProductColors = productDto.ProductColors.Split(',', StringSplitOptions.RemoveEmptyEntries)
+                                                            .Select(color => new ProductColor()
+                                                            {
+                                                                ColorName = color.Trim()
+                                                            }).ToList();
+
+
+                    foreach (var image in productDto.ProductImages)
+                    {
+                        if (image.Length > 0)
                         {
-                            product.ProductName = productDto.ProductName;
-                            product.BrandId = productDto.BrandId;
-                            product.CategoryId = productDto.CategoryId;
-                            product.ProductDescription = productDto.ProductDescription;
-                            product.PurchasePrice = productDto.PurchasePrice;
-                            product.ProfitMargin = productDto.ProfitMargin;
-                        }
-
-                        product.ProductSizes = productDto.ProductSizes.Split(',', StringSplitOptions.RemoveEmptyEntries)
-                                                              .Select(size => new ProductSize()
-                                                              {
-                                                                  SizeName = size.Trim(),
-                                                                  ProductId = product.ProductId
-                                                              }).ToList();
-                        product.ProductColors = productDto.ProductColors.Split(',', StringSplitOptions.RemoveEmptyEntries)
-                                                                .Select(color => new ProductColor()
-                                                                {
-                                                                    ColorName = color.Trim(),
-                                                                    ProductId = product.ProductId
-                                                                }).ToList();
-
-
-                        foreach (var image in productDto.ProductImages)
-                        {
-                            if (image.Length > 0)
+                            var ImageName = Path.GetFileNameWithoutExtension(Path.GetRandomFileName()) + Path.GetExtension(image.FileName);
+                            string ImageFullPath = Path.Combine(environment.WebRootPath, "uploads", ImageName);
+                            using (var stream = System.IO.File.Create(ImageFullPath))
                             {
-                                var ImageName = Path.GetFileNameWithoutExtension(Path.GetRandomFileName()) + Path.GetExtension(image.FileName);
-                                string ImageFullPath = Path.Combine(environment.WebRootPath, "uploads", ImageName);
-                                using (var stream = System.IO.File.Create(ImageFullPath))
-                                {
-                                    await image.CopyToAsync(stream);
-                                }
-                                product.ProductImages.Add(new ProductImage()
-                                {
-                                    ImageFileName = ImageName,
-                                    IsMainImage = (product.ProductImages.Count == 0)
-                                });
+                                await image.CopyToAsync(stream);
                             }
+                            product.ProductImages.Add(new ProductImage()
+                            {
+                                ImageFileName = ImageName,
+                                IsMainImage = (product.ProductImages.Count == 0)
+                            });
                         }
-                        await context.Products.AddAsync(product);
-                        await context.SaveChangesAsync();
-                        await transaction.CommitAsync();
-                        return RedirectToAction("Index");
                     }
-                    catch (Exception ex)
-                    {
-                        await transaction.RollbackAsync();
-                        return Content("Lỗi Thêm sản phẩm !");
-                    }
+                    await context.Products.AddAsync(product);
+                    await context.SaveChangesAsync();
+                    await transaction.CommitAsync();
+                    return RedirectToAction("Index");
+                }
+                catch (Exception ex)
+                {
+                    await transaction.RollbackAsync();
+                    return Content("Lỗi Thêm sản phẩm !");
                 }
             }
-            return View(productDto);
-            //var firstFile = productDto.ProductImages.First();
-            //var fileDetails = new List<string>();
-
-            //foreach (var file in productDto.ProductImages)
-            //{
-            //    if (file == firstFile) // Phần tử đầu tiên
-            //    {
-            //        fileDetails.Add($"Phần tử đầu tiên: {file.FileName}");
-            //    }
-            //    else
-            //    {
-            //        fileDetails.Add($"Phần tử khác: {file.FileName}");
-            //    }
-            //}
-
-            //// Gán dữ liệu vào ViewBag
-            //ViewBag.FileDetails = fileDetails;
-
-            //return View();
         }
         public async Task<IActionResult> Edit(int? id)
         {
@@ -171,29 +153,63 @@ namespace Project_ShoeStore_Manager.Controllers
                 product.CategoryId = productDto.CategoryId;
                 product.PurchasePrice = productDto.PurchasePrice;
                 product.ProfitMargin = productDto.ProfitMargin;
-                context.ProductSizes.RemoveRange(product.ProductSizes);
-                context.ProductColors.RemoveRange(product.ProductColors);
-                context.ProductImages.RemoveRange(product.ProductImages);
-                product.ProductSizes = productDto.ProductSizes.Split(',', StringSplitOptions.RemoveEmptyEntries)
-                                                      .Select(size => new ProductSize()
-                                                      {
-                                                          SizeName = size.Trim(),
-                                                          ProductId = product.ProductId
-                                                      }).ToList();
-                product.ProductColors = productDto.ProductColors.Split(',', StringSplitOptions.RemoveEmptyEntries)
-                                                        .Select(color => new ProductColor()
-                                                        {
-                                                            ColorName = color.Trim(),
-                                                            ProductId = product.ProductId
-                                                        }).ToList();
+
                 await context.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             return View(productDto);
         }
-        public async Task<IActionResult> Delete(int? id)
+        [HttpPost]
+        public async Task<IActionResult> Delete(int id)
         {
-
+            var product = await context.Products
+                           .Include(p => p.ProductImages)
+                           .FirstOrDefaultAsync(p => p.ProductId == id);
+            if (product == null)
+            {
+                return NotFound();
+            }
+            foreach (var pi in product.ProductImages)
+            {
+                string ImagePath = Path.Combine(environment.WebRootPath, "uploads", pi.ImageFileName);
+                System.IO.File.Delete(ImagePath);
+            }
+            product.IsDeleted = true;
+            context.ProductImages.RemoveRange(product.ProductImages);
+            await context.SaveChangesAsync();
+            return RedirectToAction("Index");
         }
     }
 }
+
+
+//context.ProductSizes.RemoveRange(product.ProductSizes);
+//context.ProductColors.RemoveRange(product.ProductColors);
+//context.ProductImages.RemoveRange(product.ProductImages);
+//product.ProductSizes = productDto.ProductSizes.Split(',', StringSplitOptions.RemoveEmptyEntries)
+//                                      .Select(size => new ProductSize()
+//                                      {
+//                                          SizeName = size.Trim()
+//                                      }).ToList();
+//product.ProductColors = productDto.ProductColors.Split(',', StringSplitOptions.RemoveEmptyEntries)
+//                                        .Select(color => new ProductColor()
+//                                        {
+//                                            ColorName = color.Trim()
+//                                        }).ToList();
+//foreach (var image in productDto.ProductImages)
+//{
+//    if (image.Length > 0)
+//    {
+//        var ImageName = Path.GetFileNameWithoutExtension(Path.GetRandomFileName()) + Path.GetExtension(image.FileName);
+//        string ImageFullPath = Path.Combine(environment.WebRootPath, "uploads", ImageName);
+//        using (var stream = System.IO.File.Create(ImageFullPath))
+//        {
+//            await image.CopyToAsync(stream);
+//        }
+//        product.ProductImages.Add(new ProductImage()
+//        {
+//            ImageFileName = ImageName,
+//            IsMainImage = (product.ProductImages.Count == 0)
+//        });
+//    }
+//}
